Move Google, ArcGis, and Facebook test utils to one utils class
Move all authentication test utils (ArcGisTestUtils, GoogleTestUtils, FacebookTestUtils) to one utils class (AuthenticationTestUtils).
This will reduce number of classes in the tests module, increase compilation time and reduce code duplication.

Also, performed next changes:
- AuthenticationUtils renamed to AuthenticationTestUtils;
- ArcGisUser / GoogleUser renamed to AccessTokenResponse;
- the utils class now is non-constructible and accessible via static methods;

Relates to: OLPEDGE-718
